chore: Fix circular dep.
