+++ src/entities/swap/swaps/v2/auraBalSwaps/index.ts
@@ -1,1 +1,2 @@
 export * from './auraBalSwaps';
+export * from './types';
